imagery_insights/samples/imagery-insights-batch-processor/requirements.txt
Flask
gunicorn
google-cloud-bigquery
google-cloud-tasks
google-cloud-aiplatform
requests
google-cloud-pubsub
python-dotenv